+++ src/DnaPSSM.cpp
@@ -290,10 +290,7 @@ string::const_iterator DnaPSSM::max_like_match(const string &target,
 		return(target.begin());
 	}
 
-	string::const_iterator max_i = target.begin() + m_max_range;
-	if(max_i > target.end() - m_chars.size()) {
-		max_i = target.end() - m_chars.size();
-	}
+	string::const_iterator max_i = target.begin() + max_offset(target.length());
 	string::const_iterator best_pos;
 	best_logp = R_NegInf;
 	for(string::const_iterator i = target.begin() + m_min_range;
@@ -374,10 +371,7 @@ void DnaPSSM::update_like_vec(const string &target,
 		return;
 	}
 
-	string::const_iterator max_i = target.begin() + m_max_range;
-	if(max_i > target.end() - m_chars.size()) {
-		max_i = target.end() - m_chars.size();
-	}
+	string::const_iterator max_i = target.begin() + max_offset(target.length());
 	vector<float>::iterator delta = deltas.begin() + m_min_range;
 	vector<float>::iterator like = likes.begin() + m_min_range;
 	vector<int1>::iterator dir = dirs.begin() + m_min_range;
@@ -502,10 +496,7 @@ void DnaPSSM::integrate_like(const string &target, float &energy, vector<float> 
 		return;
 	}
 
-	string::const_iterator max_i = target.begin() + m_max_range;
-	if(max_i > target.end() - m_chars.size()) {
-		max_i = target.end() - m_chars.size();
-	}
+	string::const_iterator max_i = target.begin() + max_offset(target.length());
 	energy = R_NegInf;
 	for(string::const_iterator i = target.begin() + m_min_range;
 	    i <= max_i;
@@ -592,10 +583,7 @@ void DnaPSSM::count_weighted(const string &target, vector<float> &wgts,
 		return;
 	}
 
-	string::const_iterator max_i = target.begin() + m_max_range;
-	if(max_i > target.end() - m_chars.size()) {
-		max_i = target.end() - m_chars.size();
-	}
+	string::const_iterator max_i = target.begin() + max_offset(target.length());
 	vector<float>::iterator wgt = wgts.begin() + m_min_range;
 	vector<int1>::iterator dir = dirs.begin() + m_min_range;
 	for(string::const_iterator i = target.begin() + m_min_range;
@@ -641,10 +629,7 @@ void DnaPSSM::count_log_weighted(const string &target, vector<float> &wgts,
 		return;
 	}
 
-	string::const_iterator max_i = target.begin() + m_max_range;
-	if(max_i > target.end() - m_chars.size()) {
-		max_i = target.end() - m_chars.size();
-	}
+	string::const_iterator max_i = target.begin() + max_offset(target.length());
 	vector<float>::iterator wgt = wgts.begin() + m_min_range;
 	vector<int1>::iterator dir = dirs.begin() + m_min_range;
 	for(string::const_iterator i = target.begin() + m_min_range;
@@ -867,10 +852,7 @@ void DnaPSSM::integrate_energy(const string &target, float &energy, vector<float
 		return;
 	}
 
-	string::const_iterator max_i = target.begin() + m_max_range;
-	if(max_i > target.end() - m_chars.size()) {
-		max_i = target.end() - m_chars.size();
-	}
+	string::const_iterator max_i = target.begin() + max_offset(target.length());
 	energy = R_NegInf;
 	int pos = 0;
 	for(string::const_iterator i = target.begin() + m_min_range;
@@ -937,10 +919,7 @@ void DnaPSSM::integrate_energy_logspat(const string &target, float &energy, vect
 		return;
 	}
 
-	string::const_iterator max_i = target.begin() + m_max_range;
-	if(max_i > target.end() - m_chars.size()) {
-		max_i = target.end() - m_chars.size();
-	}
+	string::const_iterator max_i = target.begin() + max_offset(target.length());
 	energy = R_NegInf;
 	int pos = 0;
 	for(string::const_iterator i = target.begin() + m_min_range;
@@ -1008,10 +987,7 @@ void DnaPSSM::integrate_energy_max_logspat(const string &target, float &energy, 
 		return;
 	}
 
-	string::const_iterator max_i = target.begin() + m_max_range;
-	if(max_i > target.end() - m_chars.size()) {
-		max_i = target.end() - m_chars.size();
-	}
+	string::const_iterator max_i = target.begin() + max_offset(target.length());
 
 	energy = R_NegInf;
 
@@ -1084,10 +1060,7 @@ void DnaPSSM::like_thresh_match(const string &target, float thresh,
 		return;
 	}
 
-	string::const_iterator max_i = target.begin() + m_max_range;
-	if(max_i > target.end() - m_chars.size()) {
-		max_i = target.end() - m_chars.size();
-	}
+	string::const_iterator max_i = target.begin() + max_offset(target.length());
 	for(string::const_iterator i = target.begin() + m_min_range;
 	    i <= max_i;
 	    i++) {